requirements.txt
# GUI
pyside6>=6.4.0

# Image Processing
pillow>=9.5.0
imagehash>=4.3.1
opencv-python>=4.7.0

# Metadata
exifread>=3.0.0

# Testing
pytest>=7.3.1

# Packaging
pyinstaller>=5.9.0